% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_prob}
\alias{check_prob}
\title{check probability validity}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of side probabilities}
}
\value{
True

error message if certain conditions meet
}
\description{
Creates an function to check if the probability input is validity
}
