% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_sides}
\alias{check_sides}
\title{check sides validity}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of sides}
}
\value{
True

error message if certain conditions meet
}
\description{
Creates an function to check if the sides input is validity
}
