% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{check_times}
\alias{check_times}
\title{check times validity}
\usage{
check_times(times)
}
\arguments{
\item{times}{how many times of the dice is rolled}
}
\value{
True

error message if certain conditions meet
}
\description{
Creates an function to check if the times input is validity
}
