% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\docType{data}
\name{fair_die}
\alias{fair_die}
\title{fair_die}
\format{An object of class \code{die} of length 2.}
\usage{
fair_die
}
\value{
a fair die
}
\description{
creates a fair die with sides 1~6
}
\keyword{datasets}
