% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot}
\alias{plot}
\title{plot.roll()}
\usage{
plot(x, times = 50)
}
\arguments{
\item{x}{a vector that is the output of the roll() function}
}
\value{
bar plot
# create a die and roll it 50 times
fair_die <- die()
fair_50rolls <- roll(fair_die, times = 50)

plot(fair_50rolls)
}
\description{
plot the frequencies of rolls
}
