% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(x, side = c(1, 2, 3, 4, 5, 6))
}
\arguments{
\item{x}{the die()}

\item{side}{the sides of the dice}
}
\value{
barplot
}
\description{
Plots of the relative frequencies
}
