% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot_roll}
\alias{plot_roll}
\title{plot-roll function}
\usage{
plot_roll(x)
}
\arguments{
\item{rolls}{is a set series of die rolls}
}
\value{
a barplot with "sides of die" on the x axis and "relative frequencies" on the y axis
}
\description{
returns a barplot with the frequency of each die face appearing over the number of rolls
}
