% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{+.roll}
\alias{+.roll}
\title{addition}
\usage{
\method{+}{roll}(x, incr)
}
\arguments{
\item{x}{the roll}

\item{incr}{the number of increasing toss}
}
\value{
roll with the incr numbers of toss
}
\description{
to make addition to the roll
}
