% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{print.die}
\alias{print.die}
\title{print.die}
\usage{
\method{print}{die}(x)
}
\arguments{
\item{x}{list from die()}
}
\value{
organized data frame with sides and probability
}
\description{
Creates an function to print the organized die distribution
}
