% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.roll}
\alias{print.roll}
\title{print.roll}
\usage{
\method{print}{roll}(x)
}
\arguments{
\item{x}{function that is passed in after die()}
}
\value{
a list contains rolls
}
\description{
Creates an function to print the organized data of roll()
}
