% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.summary.roll}
\alias{print.summary.roll}
\title{print.summary.roll}
\usage{
\method{print}{summary.roll}(x)
}
\arguments{
\item{x}{an organized data frame}
}
\value{
a data frame contains column of side, count, and prop
}
\description{
Creates an function to print summary for roll
}
