% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{summary.roll}
\alias{is.roll}
\title{Roll object}
\usage{
roll(die, times = 1)

\method{summary}{roll}(x, ...)

is.roll(obj)
}
\arguments{
\item{die}{object of class "die"}

\item{times}{number of times to roll the die}

\item{x}{an object \code{"roll"}}

\item{\dots}{further arguments ignored}

\item{obj}{an R object}
}
\value{
an object of class "roll"
}
\description{
main roll function
}
\examples{
# fair die
die1 <- die()

# roll a coin 5 times
roll5 <- roll(die1, times = 5)
roll100 <- roll(die(), times = 100)
summary(roll100)

}
