% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{[.roll}
\alias{[.roll}
\title{extraction}
\usage{
\method{[}{roll}(x, i)
}
\arguments{
\item{x}{the roll}

\item{i}{the index of which roll to be selected}
}
\value{
the ith index number
}
\description{
to extract a value in the roll
}
