% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-roll-methods.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replacemnt}
\usage{
\method{[}{roll}(x, i) <- value
}
\arguments{
\item{x}{the roll}

\item{i}{the index of which roll to be replaces}

\item{value}{the value that is replaced}
}
\value{
the ith index number with the replacement
}
\description{
to replace a value in the roll
}
