% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{summary.roll}
\alias{summary.roll}
\title{summary.roll}
\usage{
\method{summary}{roll}(object)
}
\arguments{
\item{object}{list of variables}
}
\value{
a data frame contains column of side, count, and prop
}
\description{
Creates an function to generate summary for roll
}
