% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{check_prob}
\alias{check_prob}
\title{check prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of probabilities}
}
\value{
error if prob no valid
}
\description{
checks that the probabilities are valid
}
\examples{
check_prob(c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))
}
