% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_prob}
\alias{check_prob}
\title{check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{probabilities of the die sides}
}
\description{
checks the probabilities to see if they are a correct input
}
