% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_prob}
\alias{check_prob}
\title{check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{a}{list of six probability values}
}
\value{
boolean
}
\description{
check if given probability adds up to 1 and is valid
}
