% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_prob}
\alias{check_prob}
\title{check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{probability}{vector}
}
\value{
TRUE or stop statement
}
\description{
checks to see if probability vector is valid
}
