% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{check_prob}
\alias{check_prob}
\title{check_prob}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{- probability of appearance of each side of a die.}
}
\value{
stops if the probability is not valid
}
\description{
check if the input probability is valid
}
