% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{check_sides}
\alias{check_sides}
\title{check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{a vector of sides to check}
}
\value{
error if sides not valid
}
\description{
checks if number of sides is valid
}
\examples{
check_sides(c(1,2,3,4,5,6))
}
