% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_sides}
\alias{check_sides}
\title{check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{sides of the die}
}
\description{
checks the number of sides of a die to see if it is a value argument
}
