% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{check_sides}
\alias{check_sides}
\title{check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of die sides}
}
\value{
TRUE if the vector sides is valid
}
\description{
check if sides is valid
}
