% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{check_sides}
\alias{check_sides}
\title{check_sides}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{- character string of a die's sides}
}
\value{
stops if the sides are not valid
}
\description{
check if the input sides is valid
}
