% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times)
}
\arguments{
\item{times}{number of rolls}
}
\value{
error if times is invalid
}
\description{
checks if times is valid
}
\examples{
check_times(50)
}
