% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times)
}
\arguments{
\item{times}{number of times the die is to be rolled}
}
\description{
aux function checks whether input is a valid number of rolls for a die
}
