% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times)
}
\arguments{
\item{times}{number of times that a die is rolled}
}
\value{
TRUE if the number times is valid
}
\description{
check if times is valid
}
