% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{check_times}
\alias{check_times}
\title{check_times}
\usage{
check_times(times)
}
\arguments{
\item{times}{number of rolls}
}
\value{
validity TRUE or FALSE
}
\description{
check the validity of the input value as a number of die roll
}
