% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))
}
\arguments{
\item{sides}{a vector of sides}

\item{prob}{a vector of probabilities for each side}
}
\value{
a dataframe of sides and probabilities
}
\description{
display information of a die
}
\examples{
# default
fair <- die()
}
