% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\alias{is.die}
\title{Die}
\usage{
die(sides = 1:6, prob = rep(1/6, 6))

is.die(x)
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class \code{"die"}
}
\description{
Creates an object of class \code{"die"}
}
\examples{
# default die
fair_die <- die()

# another die
die_1 <- die(c('h', 't'))

# loaded die
loaded <- die(prob = c(0.8, 0.04, 0.04, 0.04, 0.04, 0.04))
}
\seealso{
\code{\link{roll}}
}
