% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die object}
\usage{
die(sides = c(1:6), prob = rep(1/length(sides), length(sides)))
}
\arguments{
\item{sides}{sides of each die - six total}

\item{prob}{probabilities corresponding to each side of the die}
}
\value{
object of class "die"
}
\description{
creates an object of class "die"
}
\examples{
# default
die <- die()

# loaded die
loaded <- die(c("i", "ii", "iii", "iv", "v", "vi"),c(0.1,0.2,0.3,0.2,0.1,0.1))

}
