% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\alias{is.die}
\title{Die object}
\usage{
die(sides = 1:6, prob = rep(1/6, 6))

is.die(x)
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector with probabilities of sides}

\item{x}{an R object}
}
\value{
an object of class "die"
}
\description{
constructor function of objects of class "die"
}
\examples{
# creating a fair die (by default)
die1 <- die()

# a loaded die
loaded1 <- die(1:6, prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

}
\seealso{
\code{link{toss}}
}
