% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1:6), prob = rep(1/6, 6))
}
\arguments{
\item{a}{list of sides and list of probability}
}
\value{
'die' object
}
\description{
create an object 'die' with sides and probability
}
