% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{Die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = c(rep(1/6, 6)))
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of die probabilities}
}
\value{
an object of class \code{"die"}
}
\description{
Creates an object of class \code{"die"}
}
