% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\alias{is.die}
\title{Die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))

is.die(x)
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class \code{"die"}
}
\description{
Creates an object of class \code{"die"}
}
\examples{
# default call: creates a standard fair die
fair_die <- die()

# die with non-standard sides
weird_die <- die(sides = c('i', 'ii', 'iii', 'iv', 'v', 'vi'))

# create a loaded die
loaded_die <- die(prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
}
\seealso{
\code{\link{roll}}
}
