% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die object function}
\usage{
die(sides, prob)
}
\arguments{
\item{sides}{is a six element vector, by default numbers 1,2,3,4,5,6. must be 6 element}

\item{prob}{is a vector of probabilities for each side, by default all equal to 1/6. must sum to 1}
}
\value{
columns showing the sides of the die and the probability of rolling each one
}
\description{
returns a die object given sides and probability
}
