% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{Die}
\usage{
die(sides = c(1:6), prob = rep(1/6, 6))
}
\arguments{
\item{sides}{vector of sides}

\item{prob}{vector of probabilities}
}
\value{
printed die output
}
\description{
creates a die and prints its info
}
\examples{
# default
die1 = die()

# funky die
color_die = die(sides = c('red', 'orange', 'yellow', 'green', 'blue', 'purple'))

# loaded die
loaded = die(prob = c(0.1, 0.1, 0.1, 0.1, 0.1, 0.5))

}
