% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1:6), prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))
}
\arguments{
\item{sides}{by default numbers}

\item{prob}{probabilities}
}
\value{
data frame from html
}
\description{
create a fair die by default
}
