% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = 1:6, prob = rep(1/6, 6))
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of die probability}
}
\value{
an object with class 'die'
}
\description{
construct a die class
}
