% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_and_roll.R, R/hw5.R
\name{die}
\alias{die}
\alias{is.die}
\title{Die object}
\usage{
die(
  sides = c("1", "2", "3", "4", "5", "6"),
  prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6)
)

is.die(x)

die(
  sides = c("1", "2", "3", "4", "5", "6"),
  prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6)
)
}
\arguments{
\item{sides}{sides of a die}

\item{prob}{probability of each side to appear}

\item{x}{an R object}
}
\value{
an object of class "die"

a die with class "die"
}
\description{
constructor function of objects of class "die"

creates a valid die
}
\examples{
# creating a fair die (by default)
die1 <- die()

# a loaded die
loaded1 <- die(1:6, prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

}
\seealso{
\code{link{toss}}
}
