% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = rep(1/6, 6))
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of side probabilities}
}
\value{
an object of class "die"
}
\description{
Creates an object of class "die"
}
