% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die()}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = rep(1/6, 6))
}
\arguments{
\item{sides}{vector of six elements, by default numbers 1, 2, 3, 4, 5, 6}

\item{prob}{vector of probabilities for each side (all equal to 1/6 by default)}
}
\value{
data frame
# default call: creates a standard fair die
fair_die <- die()
# die with non-standard sides
weird_die <- die(sides = c('i', 'ii', 'iii', 'iv', 'v', 'vi'))
# create a loaded die
loaded_die <- die(prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))
}
\description{
object 'die'
}
