% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_function.R
\name{die}
\alias{die}
\title{Die}
\usage{
die(sides = seq(1, 6), prob = rep(1/6, 6))
}
\arguments{
\item{sides}{vector of die sides}

\item{prob}{vector of die probabilities}
}
\value{
an object of class die
}
\description{
Creates an object of class "die'
}
\examples{

default
fair_die <- die()

# weird die
weird_die <- die(sides = c('i', 'ii', 'iii', 'iv', 'v', 'vi'))

# loaded die
loaded_die <- die(prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

# bad die
bad_die <- die(sides = c('a', 'b', 'c', 'd', 'e'))
}
