% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6))
}
\arguments{
\item{x}{sides and probability}
}
\value{
an object
}
\description{
creating die object
}
