% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c(1, 2, 3, 4, 5, 6), prob = c(rep(1/6, 6)))
}
\arguments{
\item{sides}{a vector of sides}

\item{prob}{a vector of probs}
}
\value{
list of sides

list of prob
}
\description{
Creates an function to return the die distribution
}
