% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{die}
\alias{die}
\title{die}
\usage{
die(sides = c("1", "2", "3", "4", "5", "6"), prob = c(1/6, 1/6, 1/6, 1/6,
  1/6, 1/6))
}
\arguments{
\item{sides}{sides of a die}

\item{prob}{probability of each side to appear}
}
\value{
a die with class "die"
}
\description{
creates a valid die
}
