% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{make_roll}
\alias{make_roll}
\title{Make Roll Object}
\usage{
make_roll(side, prob, rolls, times = 1)
}
\arguments{
\item{side}{the side of a die}

\item{prob}{prob of a die}

\item{rolls}{the rolls}

\item{times}{number of rolls}
}
\description{
Constructor function for object "roll"
}
\keyword{internal}
