% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{make_roll}
\alias{make_roll}
\title{Make Roll Object}
\usage{
make_roll(die, rolls)
}
\arguments{
\item{die}{object of class die}

\item{rolls}{object of class rolls}
}
\description{
Constructor function for object "roll"
}
\keyword{internal}
