% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{make_roll}
\alias{make_roll}
\title{Make roll Object}
\usage{
make_roll(die, flips)
}
\arguments{
\item{flips}{object of class flips}

\item{coin}{object of class die}
}
\description{
Constructor function for object "roll"
}
\keyword{internal}
