% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{plot.roll}
\alias{plot.roll}
\title{plot.roll}
\usage{
\method{plot}{roll}(x)
}
\arguments{
\item{x}{output of rolls}
}
\value{
plot of rolls
}
\description{
plot the simulated die rolls
}
\examples{
fair <- die()
rollfair <- roll(fair, 50)
plot(rollfair)
}
