% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"die"}}

\item{\dots}{arguments to be passed to/from other methods}

\item{side}{number indicating which side of die to consider}
}
\description{
Plots the relative frequencies of a series of rolls
}
\examples{
 \dontrun{
 # create a die and toss it 100 times
 die1 <- die()
 roll1000 <- roll(die1, times = 1000)

 plot(roll1000)
 }
}
