% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(x, num = x$total)
}
\arguments{
\item{x}{an object of class \code{"die"}}

\item{num}{number of total rolls}
}
\description{
Plots the relative frequencies in a barplot of a series of rolls
}
