% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(x)
}
\arguments{
\item{x}{an object of class \code{"die"}}
}
\description{
Plots the relative frequencies of a series of rollss
}
\examples{
 \dontrun{
 # create a die and roll it 100 times
 die1 <- die()
 roll1000 <- roll(die1, times = 1000)

 plot(roll1000)
 }
}
