% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot Roll Objects}
\usage{
\method{plot}{roll}(roll)
}
\arguments{
\item{roll}{an object of class 'roll'}
}
\description{
plots the relative frequencies of a series of rolls
}
\examples{
# Create a die and toss it 100 times
die1 = die()
roll100 = roll(die1, times = 100)
# Plot the frequencies
plot(roll100)
}
