% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-roll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(x, side = 1)
}
\arguments{
\item{x}{an object of class \code{"die"}}

\item{side}{number indicating which side of die to consider}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the relative frequencies of a series of rolls
}
