% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{plot.roll}
\alias{plot.roll}
\title{plot.roll}
\usage{
\method{plot}{roll}(rolls)
}
\arguments{
\item{rolls}{an object of class 'roll'}
}
\value{
a plot about the roll object
}
\description{
plot the object roll
}
