% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dieroll.R
\name{plot.roll}
\alias{plot.roll}
\title{Plot of object roll}
\usage{
\method{plot}{roll}(roll, ...)
}
\arguments{
\item{roll}{an object of class \code{"roll"}}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the relative frequencies of a series of rolls
}
\examples{
 \dontrun{
 #plot method
 plot(fair_50rolls)

 #plot method
 plot(fair500, 500)

 }
}
