% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{plot.roll}
\alias{plot.roll}
\title{plot.roll}
\usage{
\method{plot}{roll}(x, ...)
}
\arguments{
\item{x}{sides and number of rolls of die}
}
\value{
a bar graph
}
\description{
shows a bar graph of the simulation of die rolls
}
