% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{plot.roll}
\alias{plot.roll}
\title{plot.roll}
\usage{
\method{plot}{roll}(x, side = 1, ...)
}
\arguments{
\item{x}{roll}
}
\value{
a plot for roll result
}
\description{
defines a plot function for roll
}
