% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{+.roll}
\alias{+.roll}
\title{addition method}
\usage{
\method{+}{roll}(x, inc)
}
\arguments{
\item{x}{roll object}

\item{inc}{amount to increment number of rolls}
}
\value{
new rolls object with increased rolls
}
\description{
adds to the number of rolls
}
\examples{
fair <- die()
rollfair <- roll(fair, 50)
morerolls <- rollfair + 30
}
