% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{+.roll}
\alias{+.roll}
\title{addition method}
\usage{
\method{+}{roll}(object, increment)
}
\arguments{
\item{rolls}{is a set series of die rolls}
}
\value{
an updated version of the roll object with the new total number of rolls
}
\description{
a function to add a specified number of rolls
}
