% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{+.roll}
\alias{+.roll}
\title{add roll}
\usage{
\method{+}{roll}(rolls, increment)
}
\arguments{
\item{rolls}{an object of class 'roll'}

\item{increment}{the number of additional rolls}
}
\value{
a new roll with new rolls included
}
\description{
add some rolls
}
