% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{+.roll}
\alias{+.roll}
\title{addition method}
\usage{
\method{+}{roll}(obj, incr)
}
\arguments{
\item{obj}{roll}

\item{incr}{number of rolls user wants to add}
}
\value{
input rolls + new rolls
}
\description{
enable adding rolls to an input roll
}
