% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.R
\name{print.die}
\alias{print.die}
\title{print.die}
\usage{
\method{print}{die}(object)
}
\arguments{
\item{object}{an object with class 'die'}
}
\value{
display the die's sides as well as its probabilities
}
\description{
print the object with class 'die'
}
