% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.roll}
\alias{print.roll}
\title{print.roll}
\usage{
\method{print}{roll}(rolls)
}
\arguments{
\item{rolls}{an object with class 'roll'}
}
\value{
a table about the roll object
}
\description{
print the roll object
}
