% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.roll}
\alias{print.roll}
\title{print.roll}
\usage{
\method{print}{roll}(die, ...)
}
\arguments{
\item{die}{and number of rolls}
}
\value{
an object in list form
}
\description{
shows the simulation results
}
