% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{print.roll}
\alias{print.roll}
\title{print.roll}
\usage{
\method{print}{roll}(x)
}
\arguments{
\item{x}{a created roll object}
}
\value{
modified print method of roll()
}
\description{
modifies the print method of roll()
}
