% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{print.summary.roll}
\alias{print.summary.roll}
\title{print.summary.roll}
\usage{
\method{print}{summary.roll}(x)
}
\arguments{
\item{x}{object}
}
\value{
output of summary
}
\description{
prints the outout of summary.roll
}
