% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.summary.roll}
\alias{print.summary.roll}
\title{print.summary.roll}
\usage{
\method{print}{summary.roll}(df)
}
\arguments{
\item{df}{an object with class 'summary.roll'}
}
\value{
a summary about the summary.roll object
}
\description{
print the summary of rolls
}
