% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{print.summary.roll}
\alias{print.summary.roll}
\title{print.summary.roll}
\usage{
\method{print}{summary.roll}(x, ...)
}
\arguments{
\item{x}{die with sides and number of roles}
}
\value{
print of summary data frame
}
\description{
shows the summary of the simulation results
}
