% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{print.summary.roll}
\alias{print.summary.roll}
\title{print.summary.roll}
\usage{
\method{print}{summary.roll}(x, ...)
}
\arguments{
\item{x}{roll}
}
\value{
modified print method of summary.roll
}
\description{
modifies the print method of summary.roll
}
