% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object (a die)}

\item{times}{number of rolls}
}
\value{
vector of die rolls
}
\description{
simulates rolling a die a given number of times
}
\examples{
fair <- die()
rollfair <- roll(fair, 50)
}
