% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object of a die}

\item{times}{the number of times rolling a die}
}
\value{
an object of class \code{"roll"}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
# default
fair_die = die()
roll1 <- roll(fair_die,50)

}
