% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolls.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(die, times)
}
\arguments{
\item{die}{object of die class}

\item{times}{number of rolls}
}
\description{
takes a die and a number of rolls, returns object class rolls
}
\examples{
# roll fair die 10 times
roll <- roll(fair_die, 10)


}
