% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a die}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object of class \code{"die"}}

\item{times}{number of rolls}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of die \code{"sides"}}

\item{prob}{vector of die \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
