% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{is.roll}
\alias{as.roll}
\title{Roll a die}
\usage{
roll(die, times = 1)

is.roll(x)

as.roll(x)
}
\arguments{
\item{die}{object of class \code{"die"}}

\item{times}{number of rolls}

\item{x}{an R object}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of die \code{"sides"}}

\item{prob}{vector of die \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
# roll fair die 50 times
fair_die <- die()
set.seed(123)
fair50 <- roll(fair_die, times = 50)

# string die
str_die <- die(
 sides = c('a', 'b', 'c', 'd', 'e', 'f'),
 prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

# roll 20 times
set.seed(123)
str_rolls <- roll(str_die, times = 20)

set.seed(123)
fair_50rolls <- roll(fair_die, times = 50)
fair50_sum <- summary(fair_50rolls)

# roll fair die
set.seed(123)
fair_die <- die()
fair500 <- roll(fair_die, times = 500)

# summary method
summary(fair500)

# extracting roll in position 500
fair500[500]

# replacing last roll
fair500[500] <- 1
fair500[500]

summary(fair500)

# adding 100 rolls
fair600 <- fair500 + 100
summary(fair600)
}
