% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{is.roll}
\alias{as.roll}
\title{Roll a die}
\usage{
roll(die, times = 1)

is.roll(x)

as.roll(x)
}
\arguments{
\item{die}{object of class \code{"die"}}

\item{times}{number of rolls}

\item{x}{an R object}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of die \code{"sides"}}

\item{prob}{vector of die \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
die1 <- die()

# roll a die 10 times
roll10 <- roll(die1, times = 10)

# add 5 more rolls
roll15 <- roll10 + 5
}
