% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{die roll function}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{is a die usually called from the other object die}

\item{times}{is a nonnegative integer}
}
\value{
the side of the die that resulted for each individual die roll
}
\description{
returns the results of rolls given the die and number of rolls
}
