% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a Die}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{an object of class die}

\item{times}{number of rolls}
}
\value{
an object of class "rolls"
}
\description{
rolls an object of class "die"
}
\examples{
die1 = die()

# Default
roll10 = roll(die1, times = 10)

# add 5 more rolls
roll15 = roll10 + 5

}
