% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a die}
\usage{
roll(die = die(), times = 1)
}
\arguments{
\item{die}{object of class \code{"die"}}

\item{times}{number of rolls}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{roll}{vector of a die roll}

\item{sides}{vector of coin \code{"sides"}}

\item{prob}{vector of coin \code{"prob"}}

\item{total}{total number of tosses}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
fair_die <- die()

#roll fair die 50 times
fair50 <- roll(fair_die, times = 50)

str_die <- die(sides = c('a', 'b', 'c', 'd', 'e', 'f'), prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

#roll 20 times
str_rolls <- roll(str_die, times = 20)
}
