% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R, R/other-roll-methods.R
\name{roll}
\alias{roll}
\alias{is.roll}
\alias{as.roll}
\title{roll die function}
\usage{
roll(die, times = 1)

is.roll(x)

as.roll(x)
}
\arguments{
\item{times}{number of rolls}

\item{x}{die object}

\item{prob}{vector of probabilities for each side of the die}

\item{x}{an R object}
}
\value{
vector of rolls
}
\description{
Simulates rolling a die given number of times
}
