% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(x, ...)
}
\arguments{
\item{die}{an object that has class 'die'}

\item{times}{number of times that a die is rolled}
}
\value{
a list that has information about the outcome of roll
}
\description{
roll a die several times
}
