% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_and_roll.R, R/hw5.R
\name{roll}
\alias{roll}
\alias{summary.roll}
\alias{is.roll}
\title{Roll object}
\usage{
roll(die = fair_die, times = 1)

\method{summary}{roll}(x)

is.roll(obj)

roll(die = fair_die, times = 1)
}
\arguments{
\item{die}{a die created by die() function}

\item{times}{number of rolls}

\item{x}{an object \code{"roll"}}

\item{obj}{an R object}

\item{\dots}{further arguments ignored}
}
\value{
an object of class "roll"

result of roll(s) that has a class "roll"
}
\description{
main roll function

rolls a die
}
\examples{
# fair die
die1 <- die()

# roll a coin 5 times
roll5 <- roll(die1, times = 5)
roll100 <- roll(die(), times = 100)
summary(roll100)

}
