% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dieroll.R
\name{roll}
\alias{roll}
\title{Roll a die}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object of class \code{"die"}}

\item{times}{number of rolls}
}
\value{
an object of class \code{"roll"} with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of die \code{"sides"}}

\item{prob}{vector of die \code{"prob"}}

\item{total}{total number of rolls}
}
\description{
Creates an object of class \code{"roll"}
}
\examples{
fair_die <- die()
set.seed(123)
fair50 <- roll(fair_die, times = 50)
fair50

# roll fair die 50 times
fair_die <- die()
# roll 50 times
set.seed(123)
fair_50rolls <- roll(fair_die, times = 50)
# print
fair_50rolls

# what's in fair50?
names(fair50)

fair50$rolls
fair50$sides
fair50$prob
fair50$total

# string die
str_die <- die(sides = c('a', 'b', 'c', 'd', 'e', 'f'), prob = c(0.075, 0.1, 0.125, 0.15, 0.20, 0.35))

# roll 20 times
set.seed(123)
str_rolls <- roll(str_die, times = 20)
names(str_rolls)
str_rolls
}
