% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll a coin}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object of class "coin"}

\item{times}{number of rolls}
}
\value{
an object of class "toss" with the following elements:

\item{rolls}{vector of rolls}

\item{sides}{vector of die "sides"}

\item{prob}{vector of die "prob"}

\item{total}{total number of rolls}
}
\description{
Creates an object of class "roll"
}
