% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{roll()}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{object of class 'die'}

\item{times}{number of times to roll the die (default value of 1)}
}
\value{
rolls: vector with outputs of the rolls; sides: vector with the sides of the die; prob: vector with probabilities for each side of the die
# roll fair die 50 times
fair_die <- die()

# roll 50 times
fair_50rolls <- roll(fair_die, times = 50)
# add 10 more rolls
fair_60rolls <- fair_50rolls + 10
}
\description{
object 'roll'
}
