% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_function.R
\name{roll}
\alias{roll}
\alias{is.roll}
\title{Roll}
\usage{
roll(x, times)

is.roll(x)
}
\arguments{
\item{x}{an R object}

\item{times}{number of times rolled}

\item{die}{object of class 'die'}
}
\value{
an object of class roll

\item{rolls}{vector of roll outputs}

\item{sides}{vector with sides of die}

\item{prob}{vector of probabilities for each side of the die}

\item{total}{total number of tosses}

\item{total}{total number of rolls = times}
}
\description{
Creates an object of class 'roll'
}
\examples{
fair rolls
set.seed(123)
fair_die <- die()
fair_50rolls <- roll(fair_die, times=50)
}
