% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(die, times = 1)
}
\arguments{
\item{die}{(after passing through die())}

\item{times}{how many times the dice is rolled}
}
\value{
a list contains rolls, sides, prob, and total
}
\description{
Creates an function to generate random samples
}
