% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{roll}
\alias{roll}
\title{roll}
\usage{
roll(die = fair_die, times = 1)
}
\arguments{
\item{die}{a die created by die() function}

\item{times}{number of rolls}
}
\value{
result of roll(s) that has a class "roll"
}
\description{
rolls a die
}
