% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{[.roll}
\alias{[.roll}
\title{extraction method}
\usage{
\method{[}{roll}(x, i)
}
\arguments{
\item{rolls}{is a set series of die rolls}
}
\value{
the value of a specified roll
}
\description{
a function to extract the value of a given roll
}
