% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{[.roll}
\alias{[.roll}
\title{extraction function}
\usage{
\method{[}{roll}(x, i)
}
\arguments{
\item{x}{roll}

\item{i}{index number}
}
\value{
extraction enabled
}
\description{
enable extracting an element from roll
}
