% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{[.roll}
\alias{[.roll}
\title{extraction}
\usage{
\method{[}{roll}(roll, i)
}
\arguments{
\item{x, }{sides and rolls of a die}
}
\value{
value of the i th roll
}
\description{
extract the value of a given roll
}
