% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replacement method}
\usage{
\method{[}{roll}(x, i) <- value
}
\arguments{
\item{x}{roll object}

\item{i}{index}

\item{value}{replacement value}
}
\value{
new value at index i
}
\description{
replaces value at index i with value
}
\examples{
fair <- die()
rollfair <- roll(fair, 50)
rollfair[35] <- 2
}
