% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-methods.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replacement method}
\usage{
\method{[}{roll}(x, i) <- value
}
\arguments{
\item{rolls}{is a set series of die rolls}
}
\value{
an updated version of the number of rolls
}
\description{
a function to replace the value of a given roll
}
