% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replace roll}
\usage{
\method{[}{roll}(rolls, i) <- value
}
\arguments{
\item{rolls}{an object of class 'roll'}

\item{i}{the index of roll that will be replaced}

\item{value}{the value that will replace the ith entry in the roll}
}
\value{
a new roll with ith entry replaced by value
}
\description{
replace a specific roll
}
