% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replacement}
\usage{
\method{[}{roll}(roll, i) <- value
}
\arguments{
\item{x, }{sides and rolls of a die}
}
\value{
roll object
}
\description{
replace the value of a given roll
}
