% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{[<-.roll}
\alias{[<-.roll}
\title{replacement method}
\usage{
\method{[}{roll}(x, i) <- value
}
\arguments{
\item{x}{roll}

\item{i}{index number}
}
\value{
replacement enabled
}
\description{
enable replacing an element of roll
}
