% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolldie.R
\name{summary.roll}
\alias{summary.roll}
\title{summary.roll}
\usage{
\method{summary}{roll}(x)
}
\arguments{
\item{x}{output of roll}
}
\value{
summary of roll
}
\description{
provides a table of counts and proportions for each side
}
\examples{
fair <- die()
rollfair <- roll(fair, 50)
summary(rollfair)
}
