% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{summary.roll}
\alias{summary.roll}
\title{Summary.roll}
\usage{
\method{summary}{roll}(x)
}
\arguments{
\item{x}{an object of class 'roll'}
}
\value{
an object of class \code{"summary.roll"}
}
\description{
Creates an object of class \code{"summary.roll"}
}
\examples{
# default
fair_die = die()
roll1 <- roll(fair_die,50)
roll_sum = summary(roll1)

}
