% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-roll.R
\name{summary.roll}
\alias{summary.roll}
\title{summary-roll function}
\usage{
\method{summary}{roll}(x, ...)
}
\arguments{
\item{die}{is a die usually called from the other object die}
}
\value{
a dataframe with the side, count, and proportion of each roll of a die
}
\description{
returns a dataframe with the side, count, and proportion of each roll of a die
}
