% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{summary.roll}
\alias{summary.roll}
\title{summary.roll}
\usage{
\method{summary}{roll}(rolls)
}
\arguments{
\item{rolls}{an object with class 'roll'}
}
\value{
a summary about the roll object
}
\description{
summarize the roll object
}
