% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{summary.roll}
\alias{summary.roll}
\title{summary.roll}
\usage{
\method{summary}{roll}(x, ...)
}
\arguments{
\item{x}{a certain die}
}
\value{
data frame of sides, count, and proportions
}
\description{
gives a summary of the simulated rolls
}
