% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die_object.R
\name{summary.roll}
\alias{summary.roll}
\title{summary.roll}
\usage{
\method{summary}{roll}(x)
}
\arguments{
\item{x}{a created roll object}
}
\value{
modified summary method
}
\description{
modifies the summary method of a roll
}
